Extract Donation Context
This extracts the Donation Context and most of the Payment Context
into https://github.com/wmde/fundraising-donations

A bunch of infrastructure code and random stuff also had to be
extracted as it was dependend upon by the former code. Some of
this should be put in a nicer place if we continue with this
approach, and some of it should be fixed dependency wise regardless
if we continue or not (though depends on that decision for how to).

This code can be found at https://github.com/wmde/fundraising-donations/tree/master/todo

All files where just moved and not altered. This includes no changes
in their namespace. Namespace changes likely make sense though if
done at present would get in the way of easy revert.

There is one exception: IbanValidator was renamed to KontoCheckIbanValidator
and now implements a new IbanValidator intetface defined by the new component.
This was needed to avoid coupling to the evil konto check stuff in the new component.