Refactor confirmation state
The LegacyConverter code made it clear that the "confirmed" status is
dependent on the payment, if it was booked or not. This led to an
inconsistency where a Direct Debit membership was unconfirmed by default
and confirmed after the database round trip.

This change removes the `confirm` method and uses the payment method and
state to determine the confirmation state. Non-Bookable payments (e.g.
direct debit and bank transfer) are auto-confirmed.