Check interval in Membership validation
The validation in ApplyForMembership ignored the interval error from the
ValidateMembershipFee use case it used internally. This change checks
the different results from the fee validation (which includes the
interval).

Add @covers tags to tests to cover value objects that are tested
implicitly.

This change is a reaction to an error that occurred in our error log.