Make allowed payment types configurable
Change MembershipPaymentValidator to check a list of allowed payment
types instead of hard-coding them. The list will be passed from an upper
layer (Fundraising Application) into to PaymentServiceFactory

Remove accidental duplication of MembershipNotifier interface and
implementation. This happened because we needed similar functionality
for the moderation email like we did during the payment integration. The
branches were developed in parallel.

This is a preparation for https://phabricator.wikimedia.org/T312087