Make membership id non-nullable
Use `MembershipIdGenerator` to make sure there is always an ID.
This gets rid of `hasId` and `assignId`

Simplify `DoctrineApplicationRepository`, removing branch for
creating/updating membership applications.

Rename `getApplicationById` to `getUnexportedMembershipApplicationById` to
make it more explicit what this special method does.

Remove the "explorative" test `MembershipApplicationInsertionTest` - it
was just testing if Doctrine works.

This change breaks backwards-compatibility of `ApplyForMembershipUseCase`
because it introduces another dependency.

https://phabricator.wikimedia.org/T344414