Create MembershipAuthorizer
The MembershipAuthorizer will be implemented by the fundraising app.
Add MembershipAuthorizer as a dependency.
Use new DomainSpecificPaymentCreationRequest to pass more information to
the payment.

Change ApplyForMembershipResponse to remove the tokens - they should not
be returned, but gotten in the code that needs them, probably the same
class that implements MembershipAuthorizer.

Remove all TokenFetcher dependencies from CreatePaymentUseCase and ShowApplicationConfirmationUseCase

https://phabricator.wikimedia.org/T343846