Update Symfony to 7.0
Update all Symfony dependencies to 7.0. These are only dev dependencies.
We allow for 6.x version of the cache to ensure compatibility with an
un-updated payments version.
Bump minimum PHP version to 7.0, the requirement for Symfony 7

Checked the CHANGELOG files of each dependency and we don't have any
code that is affected by breaking changes