Remove external coverage annotation
Removed `WMDE\FunValidators\ValidationResult` as a covered class from
MembershipApplicationValidatorTest.

The PHPUnit warning at this point is correct, since coverage for a
project should only be collected for the source code of that project and
not for external dependencies. The coverage tag was probably an artifact
from extracting the bounded contexts out of the Fundraising Application.