Add AnonymousEmailAddress
The EmailAddress class validates that the string passed is
in the format of an email. The email is cleared once a
membership request is anonymised meaning this check throws
validation errors.

EmailAddress doesn't have an interface so this uses inheritance
to add a new AnonymousEmailAddress child class that can be used
in place of it.

Ticket: https://phabricator.wikimedia.org/T316201