Allow export memberships to be fetched
The membership confirmation page is now able to handle
anonymised membership applications so we no longer need
to show a separate page.

This adds a method to allow fetching an exported
membership to the MembershipRepository and calls that
in the ShowApplicationConfirmationUseCase.

It also adds/adjusts tests for the above.

Ticket: https://phabricator.wikimedia.org/T316201