Add backup class for personal data
Before scrubbing the personal data from the database, we want to write
it to an encrypted short-term storage, to be able to restore it in cases
when scrubbing goes awry.

This code introduces a class, PersonalDataBackup that calls a
DatabaseBackupClient interface for writing the data and then marks the
exported data with a date. The actual implementation of the
DatabaseBackupClient will be done by the Fundraising Operation Center.

Ticket: https://phabricator.wikimedia.org/T372790