Add isScrubbed flag to database and use it
This makes the DoctrineMembershipAnonymizer faster because it won't
scrub all memberships but only the un-scrubbed ones.

The previous scrubbing method, `anonymizeAt` relied on the "backup
date", but we should no longer misuse that, backing up and scrubbing
should be two different things (to enable scrubbing in test systems
where we don't do backups) and which relied on the business logic in an
external shell script to mark the "right" data sets as backed up.

Ticket: https://phabricator.wikimedia.org/T409090